refactor(ui): decompose app.tsx god component into AppShell + useAppOrchestration (b1)
The 796-line `App` mixed auth/setup gating, ~580 lines of runtime wiring (theme,
settings, interfaces, cards, SSE, polling, run-all-tests orchestration), and the
entire authenticated render. Split it three ways with no behaviour change:

- `app/useAppOrchestration.ts` — all non-gating hooks/state/effects, returning a
  typed bundle. Called UNCONDITIONALLY in `App` (before any gating return) so
  hook order, effect timing, and global side effects (theme on the login screen,
  logger auth toggle) are byte-for-byte identical.
- `app/AppShell.tsx` — pure presentation: builds the AppContext value + top bar
  and renders router/sidebar/routes/drawers/FAB/palette from the bundle.
- `app.tsx` — now 102 lines: gating (setup wizard / loading / login) + session
  expiry + mounting <AppShell>.

No new state introduced. tsc + Biome clean; app.test.tsx (13) green; production
build green. SEED_UI_ARCH_PLAN.md B1.
